chore: add package details to mix.exs file (#7)
## Checklist

<!--
For each bullet, ensure your pr meets the criteria and write a note
explaining how this PR relates. Mark them as complete as they are done.
All top-level checkboxes should be checked regardless of their relevance
to the pr with a note explaining whether they are relevant or not.
-->

- [X] Code conforms to the [Elixir
Styleguide](https://github.com/christopheradams/elixir_style_guide)

## Problem

This is required for common-config-elixir to add package release stuff

## Details

I did not set the org, so this package will be public. We should include
a license as well @doomspork